Should be able to pass in name to hello world
This is a squash rebase of 18 commits I made in the process of getting the right node module syntax and URL module. I had to make a new commit for every little tweak because my cycles were constrained by creating Terraform plans to deploy the app. This node app is (obviously) just a toy in the service of demonstrating some Terraform stuff I'm learning.